+++ emcache/__init__.py
@@ -1,21 +1,25 @@
-from .base_client import Client, Item
+from .base import Client, ClusterEvents, Item
 from .client import create_client
-from .client_errors import NotStoredStorageCommandError, StorageCommandError
-from .cluster import MemcachedHostAddress
+from .client_errors import ClusterNoAvailableNodes, NotStoredStorageCommandError, StorageCommandError
 from .default_values import (
     DEFAULT_CONNECTION_TIMEOUT,
     DEFAULT_MAX_CONNECTIONS,
+    DEFAULT_PURGE_UNHEALTHY_NODES,
     DEFAULT_PURGE_UNUSED_CONNECTIONS_AFTER,
     DEFAULT_TIMEOUT,
 )
+from .node import MemcachedHostAddress
 
 __all__ = (
     "Client",
+    "ClusterEvents",
+    "ClusterNoAvailableNodes",
     "create_client",
     "DEFAULT_TIMEOUT",
     "DEFAULT_CONNECTION_TIMEOUT",
     "DEFAULT_MAX_CONNECTIONS",
     "DEFAULT_PURGE_UNUSED_CONNECTIONS_AFTER",
+    "DEFAULT_PURGE_UNHEALTHY_NODES",
     "Item",
     "MemcachedHostAddress",
     "NotStoredStorageCommandError",